+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-components-form",
-  "version": "1.4.0",
+  "version": "1.4.1",
   "description": "React form components",
   "main": "main.js",
   "jsnext:main": "src/index.js",
