+++ src/components/styled/Bootstrap.js
@@ -1,31 +1,10 @@
 import FieldRestyle from '../FieldsRestyle';
-import TextField from '../TextField';
-import DateField from '../DateField';
-import TextareaField from '../TextareaField';
-import NumberField from '../NumberField';
-import SubmitField from '../SubmitField';
-import SelectField from '../SelectField';
-import CheckboxField from '../CheckboxField';
-import ObjectField from '../ObjectField';
-import ListField from '../ListField';
-
-const fields = {
-    TextField,
-    DateField,
-    TextareaField,
-    NumberField,
-    SubmitField,
-    SelectField,
-    CheckboxField,
-    ObjectField,
-    ListField
-};
 
 const errorStyles = {
     className: 'alert alert-danger'
 };
 
-const Fields = FieldRestyle(fields, {
+const Fields = FieldRestyle({
     TextField: {
         className: 'form-control',
         errorStyles
